+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## [1.0.15] - 2017-05-30
 ### fixed
-- bug reading config file for aws setting
-- bug reading config file for user and password
+- bug reading config file for aws setting, user, password and indexname
 ### added
 - elasticsearch and requests to requirements.txt for pip install